Simplify loading of classic games

+++ Model/Core/GameInitializationConfig.cs
@@ -48,7 +48,7 @@ namespace Civ2engine
         public string ScenarioName { get; set; }
         public int ScenPlayerCivId { get; set; }
         public string[] CivNames { get; set; }
-        public byte[] CivGenders { get; set; }
+        public int[] CivGenders { get; set; }
         public string[] LeaderNames { get; set; }
         public bool[] CivsInPlay { get; set; }
         public string LeaderName { get; set; }
